yarnでのscripts追加 ファイル名の修正など
