+++ package.json
@@ -38,6 +38,7 @@
     "clear:cache": "npx shx rm -rf node_module/.cache dist bundles test-dist",
     "lint": "eslint {app,test}/**/*.{ts,js,vue} && stylelint app/**/*.{vue,css,less}",
     "format": "prettier -w {app,test}/**/*.{ts,js,vue} && eslint --fix {app,test}/**/*.{ts,js,vue} && stylelint --fix app/**/*.{vue,css,less} && npx sort-package-json",
+    "update-gamecapture": "node bin/update-gamecapture.js",
     "postinstall": "node additional/main.mjs"
   },
   "husky": {
